requirements.txt
streamlit==1.29.0
pycryptodome==3.19.0
python-dotenv==1.0.0
pillow==10.1.0